move metrics.rs to execution mod
