+++ scylla/src/execution/mod.rs
@@ -2,6 +2,7 @@ mod execution_profile;
 pub mod history;
 pub mod iterator;
 pub mod load_balancing;
+pub(crate) mod metrics;
 pub mod retries;
 pub mod speculative_execution;
 pub mod tracing;
@@ -11,3 +12,5 @@ pub use execution_profile::{ExecutionProfile, ExecutionProfileBuilder, Execution
 
 #[cfg(test)]
 pub(crate) use execution_profile::defaults;
+
+pub use metrics::Metrics;
