+++ scylla/src/lib.rs
@@ -138,5 +138,3 @@ pub use session::CloudSessionBuilder;
 
 pub use cluster::host_filter;
 pub use execution::{load_balancing, speculative_execution};
-
-pub use transport::metrics::Metrics;
