+++ scylla/src/transport/mod.rs
@@ -1,4 +1,3 @@
-pub(crate) mod metrics;
 pub mod partitioner;
 pub mod query_result;
